backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jira==3.5.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
plotly==5.17.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
# Capacity Planning Dependencies
numpy==1.24.3
reportlab==4.0.4
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2 